% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{aft_pec}
\alias{aft_pec}
\title{Calculate prediction error curve}
\usage{
aft_pec(
  object,
  ipcw.form = Surv(time = start, time2 = end, event = event) ~ 1,
  newdata,
  times,
  ...
)
}
\arguments{
\item{object}{A fitted model from which to extract predicted survival
probabilities}

\item{ipcw.form}{The formula used to calculate inverse probability censoring
weights.}

\item{newdata}{A data frame containing predictor variable combinations for
which to compute predicted survival probabilities.}

\item{times}{A vector of times in the range of the response variable, e.g.
times when the response is a survival object, at which to return the
survival probabilities.}

\item{...}{Further arguments passed to \code{\link[pec]{predictSurvProb}}}
}
\description{
Calculate prediction error curve
}
\examples{
library(eha)
data(mort)
mort_train <- mort[1:1000,]
mort_test <- mort[-1:-1000,]
mod <- aftreg(Surv(enter, exit, event) ~ ses, param = "lifeExp",
 data = mort_train)
pec <- aft_pec(
 mod,
 ipcw.form = Surv(time = enter, time2=exit, event = event) ~ ses,
 times = seq(0, max(mort_test$exit) - 1, by = 1),
 newdata = mort_test)
plot(seq(0, max(mort_test$exit)-1, by = 1), pec, type = "l")
}
