% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{aft_pec_cv}
\alias{aft_pec_cv}
\title{Calculate cross-validated prediction error curves}
\usage{
aft_pec_cv(
  data,
  mod.form,
  ipcw.form,
  times = c(0, 30, 100),
  seed = NULL,
  k = 10,
  ...
)
}
\arguments{
\item{data}{The data set used for cross-validated brier-score calculation.}

\item{mod.form}{The model specification used to fit the model in each fold.
Will be passed to \code{\link[eha]{aftreg}}.}

\item{ipcw.form}{The specification of the model that calculates the
inverse probability censoring weights (ipcw). Will be passed to
\code{\link[pec]{ipcw}}.}

\item{times}{For \code{what="IPCW.times"} a vector of times at
which to compute the probabilities of not being censored.}

\item{seed}{If specified this seed will be used (to replicate CV folds for example).}

\item{k}{Number of folds (an integer).}

\item{...}{Further arguments passed to \code{\link{aft_pec}}.}
}
\description{
Calculate cross-validated prediction error curves
}
\examples{
library(eha)
pec_cv_df <- aft_pec_cv(
 data = mort,
 mod.form = "Surv(time = enter, time2 = exit, event = event)~ses",
 ipcw.form = "Surv(time = enter, time2 = exit, event = event)~ses",
 times = seq(0, 10, by = 1),
 method = "cox",
 k = 5)
pec_cv_df <- aft_pec_cv(
 data = mort,
 mod.form = "Surv(time = enter, time2 = exit, event = event)~ses",
 ipcw.form = "Surv(time = enter, time2 = exit, event = event)~ses",
 times = seq(0, 10, by = 1),
 method = "aftreg",
 k = 5)
}
