% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{aft_pec_wrapper}
\alias{aft_pec_wrapper}
\title{Wrapper functions that calculates CV prediction error}
\usage{
aft_pec_wrapper(
  data,
  num_vars,
  cat_vars,
  ipcw.form,
  times,
  lhs = "Surv(time = start, time2 = end, event = event)~",
  seed = NULL,
  k = 10,
  fname = NULL,
  reference = TRUE,
  ...
)
}
\arguments{
\item{data}{The data set used for cross-validated brier-score calculation.}

\item{num_vars}{A character vector of covariates on numeric scale.}

\item{cat_vars}{A character vector of categorical covariates}

\item{ipcw.form}{The specification of the model that calculates the
inverse probability censoring weights (ipcw). Will be passed to
\code{\link[pec]{ipcw}}.}

\item{times}{For \code{what="IPCW.times"} a vector of times at
which to compute the probabilities of not being censored.}

\item{lhs}{The left-hand side of the formula to be produced.}

\item{seed}{If specified this seed will be used (to replicate CV folds for example).}

\item{k}{Number of folds (an integer).}

\item{fname}{A string of the special function in which \code{x} will be
wrapped in. Defaults to \code{"rcs"}}

\item{reference}{Logical. If \code{TRUE} (default) a baseline model
(\code{ ~ 1}) is added for comparison.}

\item{...}{Further arguments passed to \code{\link{aft_pec}}.}
}
\description{
Wrapper functions that calculates CV prediction error
}
\examples{
library(eha)
pec <- aft_pec_wrapper(
 data = mort,
 num_vars = "ses",
 ipcw.form = "Surv(time = enter, time2 = exit, event = event)~ses",
 times = seq(0, 17, by = 1),
 k = 5,
 lhs = "Surv(time = enter, time2 = exit, event = event)~",
 fname = "")
}
