% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-parser.R
\name{create_formulas}
\alias{create_formulas}
\title{Create formulas for all possible combinations of variables}
\usage{
create_formulas(num_vars, cat_vars, lhs, fname = NULL)
}
\arguments{
\item{num_vars}{A character vector of covariates on numeric scale.}

\item{cat_vars}{A character vector of categorical covariates}

\item{lhs}{The left-hand side of the formula to be produced.}

\item{fname}{A string of the special function in which \code{x} will be
wrapped in. Defaults to \code{"rcs"}}
}
\value{
A vector of formula strings for possible (and correct) model
specifications given the variables provided.
}
\description{
Creates formulas for all possible combinations of variables provided to
the function. All terms can appear as linear terms or non-linear terms.
}
\examples{
create_formulas(num_vars = c("x1", "x2"), lhs = "Surv(time, status) ~")
create_formulas(num_vars = c("x1", "x2"), cat_vars = c("z1"),
 lhs = "Surv(time, status) ~")
create_formulas(num_vars = c("x1"), cat_vars = c("z1", "pspline(x2)"),
 lhs = "Surv(time, status) ~")
create_formulas(num_vars = c("x1", "x2"), cat_vars = c("z1"),
 lhs = "Surv(time, status) ~", fname = "psline")
}
