% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-parser.R
\name{fpaste}
\alias{fpaste}
\title{Wrapes variable in a function}
\usage{
fpaste(x, fname = "rcs")
}
\arguments{
\item{x}{A String the will be wrapped in \code{fname}.}

\item{fname}{A string of the special function in which \code{x} will be
wrapped in. Defaults to \code{"rcs"}}
}
\description{
Wrapes variable in a function
}
\seealso{
\code{\link[rms]{rms.trans}}
}
\keyword{internal}
