% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{get_best}
\alias{get_best}
\title{Extract formula of the best model}
\usage{
get_best(pec_df, times = NULL)
}
\arguments{
\item{pec_df}{A data frame containing prediction error (brier score)
on different time points.}

\item{times}{The time(s) for which the best fit will be searched.}
}
\description{
Extract formula of the best model
}
