% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{get_best_form}
\alias{get_best_form}
\title{Extract specification of model with the lowest brier score}
\usage{
get_best_form(pec_df, times = NULL)
}
\arguments{
\item{pec_df}{A data frame containing prediction error (brier score)
on different time points.}

\item{times}{The time(s) for which the best fit will be searched.}
}
\value{
The formula of the model that produced the lowest brier score at
time \code{time}.
}
\description{
Extract specification of model with the lowest brier score
}
