% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_utility.R
\name{get_coxsnell}
\alias{get_coxsnell}
\title{Extract data needed to create Cox-Snell vs. Cumulative hazard plot}
\usage{
get_coxsnell(fit)
}
\arguments{
\item{fit}{
    the result of fitting a Cox regression model, using the
    \code{coxph} or \code{coxme} functions. 
  }
}
\value{
A data frame containing (not censored) Cox-Snell residuals
(\code{coxsnell}) and Nelson-Aalen estimate of the Cumulative Hazard
(\code{Lcumu_hazard}) as well as the Breslow estimate for the Survival
function (\code{survival}) for censored data.
}
\description{
Extract data needed to create Cox-Snell vs. Cumulative hazard plot
}
\examples{
library(survival)
data("tongue", package="KMsurv")
cox.tongue <- coxph(Surv(time, delta)~as.factor(type), data=tongue)
cs.data <- get_coxsnell(cox.tongue)
head(cs.data)
}
