% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_utility.R
\name{get_csdata}
\alias{get_csdata}
\title{Extract data needed to create Cox-Snell vs. Cumulative hazard plot}
\usage{
get_csdata(object)
}
\arguments{
\item{object}{An object of class \code{\link[survival]{coxph}}.}
}
\value{
A data frame containing (not censored) Cox-Snell residuals (\code{coxsnell})
and Nelson-Aalen estimate of the Cumulative Hazard (\code{H}) as well as the
Breslow estimate for the Survival function (\code{S}) for censored data.
}
\description{
Extract data needed to create Cox-Snell vs. Cumulative hazard plot
}
\examples{
library(survival)
data("tongue", package="KMsurv")
cox.tongue <- coxph(Surv(time, delta)~as.factor(type), data=tongue)
cs.data <- get_csdata(cox.tongue)
head(cs.data)
}

