% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-parser.R
\name{get_multappear}
\alias{get_multappear}
\title{Check if string appears in character vector multiple times}
\usage{
get_multappear(char, vec)
}
\arguments{
\item{char}{The string for which to check multiple appearances in \code{vec}.}

\item{vec}{The character vector in which to check multiple appearences of \code{char}.}
}
\value{
\code{TRUE} if \code{char} appears more than once in \code{vec},
\code{FALSE} otherwise.
}
\description{
Check if string appears in character vector multiple times
}
