% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_utility.R
\name{get_scaledsch}
\alias{get_scaledsch}
\title{Extract scaled Schoenfeld residuals.}
\usage{
get_scaledsch(fit, transform = "km")
}
\arguments{
\item{fit}{
    the result of fitting a Cox regression model, using the
    \code{coxph} or \code{coxme} functions. 
  }

\item{transform}{
    a character string specifying how the survival times should be transformed 
    before the test is performed. 
    Possible values are \code{"km"}, \code{"rank"}, \code{"identity"} or a 
    function of one argument. 
  }
}
\value{
A tidy data frame containing the (transformed) time and scaled
Schoenfeld residuals for the variables used in \code{fit}
}
\description{
Extract scaled Schoenfeld residuals from
\code{\link[survival]{coxph}} object in tidy format.
}
