% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_utility.R
\name{get_term}
\alias{get_term}
\title{Extract partial effects for specified model terms}
\usage{
get_term(fit, data, term, ...)
}
\arguments{
\item{fit}{
    the result of fitting a Cox regression model, using the
    \code{coxph} or \code{coxme} functions. 
  }

\item{data}{Any data frame containing variables used to fit the model. Only
first row will be used.}

\item{term}{The (non-linear) model term of interest.}

\item{...}{Currently ignored.}
}
\description{
Extract partial effects for specified model terms
}
\keyword{internal}
