% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_utility.R
\name{get_terms}
\alias{get_terms}
\title{Extract the partial effects of non-linear model terms}
\usage{
get_terms(fit, data, terms, ...)
}
\arguments{
\item{fit}{
    the result of fitting a Cox regression model, using the
    \code{coxph} or \code{coxme} functions. 
  }

\item{data}{Any data frame containing variables used to fit the model. Only
first row will be used.}

\item{terms}{A character vector (can be length one). Specifies the terms
for which partial effects will be returned}

\item{...}{Currently ignored.}
}
\value{
A data frame with 5 columns.
}
\description{
Extract the partial effects of non-linear model terms
}
\examples{
library(survival)
fit <- coxph(Surv(time, status) ~ pspline(karno, df=4), data=veteran)
term.karno <- get_terms(fit, veteran, terms="karno")
}
\seealso{
\code{\link[survival]{coxph}}
}
