% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_plots.R
\name{gg_coxsnell}
\alias{gg_coxsnell}
\title{Overall model fit diagnostic via Cox-Snell residuals.}
\usage{
gg_coxsnell(fit, type = c("cumu_hazard", "cdf"))
}
\arguments{
\item{fit}{
    the result of fitting a Cox regression model, using the
    \code{coxph} or \code{coxme} functions. 
  }

\item{type}{\code{character}. Optional argument. If \code{"cumu_hazard"} plots
cumulative hazard (Nelson-Aalen estimate) vs. Cox-Snell residuals.
If \code{"cdf"} plot empirical cumulative distribution Function (Breslow estimate)
vs. Cox-Snell residuals.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
\code{gg_coxsnell} extracts Cox-Snell residuals from an \code{\link[survival]{coxph}}
object (which for a correct model should be a censored sample from Exp(1)) and
the cumulative hazard rate of the residuals vs. the residuals.
}
\examples{
library(survival)
library(ggplot2)
data("tongue", package="KMsurv")
cox.tongue <- coxph(Surv(time, delta)~as.factor(type), data=tongue)
gg_coxsnell(cox.tongue) +
 geom_abline(intercept=0, slope=1, col=2)
gg_coxsnell(cox.tongue, type="cdf") +
 geom_line(aes(y=F), col=2)
}
