% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_plots.R
\name{gg_scaledsch}
\alias{gg_scaledsch}
\title{Plot scaled schoenfeld residuals vs. time}
\usage{
gg_scaledsch(fit, transform = "identity")
}
\arguments{
\item{fit}{
    the result of fitting a Cox regression model, using the \code{coxph} function. 
  }

\item{transform}{
    a character string specifying how the survival times should be transformed 
    before the test is performed. 
    Possible values are \code{"km"}, \code{"rank"}, \code{"identity"} or  a 
    function of one argument. 
  }
}
\description{
Creates a \code{ggplot} object/plot of scaled schoenfeld residuals 
vs. (transformed) time.
}

