% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_info.R
\name{int_info}
\alias{int_info}
\alias{int_info.default}
\title{Create start/end times and interval information}
\usage{
int_info(x, ...)

\method{int_info}{default}(x, right_closed = TRUE, min_time = 0L, ...)
}
\arguments{
\item{x}{A numeric vector of cut points in which the follow-up should be
partitioned in or object of class \code{ped}.}

\item{...}{Currently ignored.}

\item{right_closed}{Logical. If \code{TRUE} (default), intervals are assumed right_closed
closed and left open. If \code{FALSE} left closed and right_closed open.}

\item{min_time}{Only intervals that have lower borders larger than
this value will be included in the resulting data frame.}
}
\value{
A data frame containing the start and end times of the
intervals specified by the \code{x} argument. Additionally the interval
length, interval mid-point and a factor variable of the intervals themselves.
}
\description{
Given interval breaks points, returns data frame with information on
interval start time, interval end time, interval length and a interval factor
variable (left open intervals). If object of class ped is provided, extracts
unique interval information from object.
}
\examples{
## create interval information from cut points
int_info(c(1, 2.3, 5))

}
