% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{integrated_brier}
\alias{integrated_brier}
\title{Calculate integrated brier score}
\usage{
integrated_brier(pec_df)
}
\arguments{
\item{pec_df}{A data frame containing prediction error (brier score)
on different time points.}
}
\value{
A data frame containing the integrated brier socre (IBS) for
each id. Results are sorted with respect to lowest IBS.
}
\description{
Calculate integrated brier score
}
