% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{ipcw.aftreg}
\alias{ipcw.aftreg}
\title{Inverse probability censoring weights (IPCW) for left-truncated data}
\usage{
\method{ipcw}{aftreg}(
  formula,
  data,
  method = "aftreg",
  args,
  times,
  subjectTimes,
  subjectTimesLag,
  what = "IPCW.times"
)
}
\arguments{
\item{formula}{A survival formula like, \code{Surv(time,status)~1}, where
as usual status=0 means censored. The status variable is internally
reversed for estimation of censoring rather than survival
probabilities. Some of the available models (see argument
\code{model}) will use predictors on the right hand side of the
formula.}

\item{data}{The data used for fitting the censoring model}

\item{method}{Censoring model used for estimation of the
(conditional) censoring distribution.}

\item{args}{A list of arguments which is passed to method}

\item{times}{For \code{what="IPCW.times"} a vector of times at
which to compute the probabilities of not being censored.}

\item{subjectTimes}{For \code{what="IPCW.subjectTimes"} a vector of
individual times at which the probabilities of not being censored
are computed.}

\item{subjectTimesLag}{If equal to \code{1} then obtain
\code{G(T_i-|X_i)}, if equal to \code{0} estimate the conditional
censoring distribution at the subjectTimes,
i.e. (\code{G(T_i|X_i)}).}

\item{what}{Decide about what to do: If equal to
\code{"IPCW.times"} then weights are estimated at given
\code{times}.  If equal to \code{"IPCW.subjectTimes"} then weights
are estimated at individual \code{subjectTimes}.  If missing then
produce both.}
}
\description{
This function uses \code{eha::aftreg} to estimate IPCW for left-truncated
data based on an AFT Weibull model. Initially designed to be used in
conjunction with \code{pec::pec}, however, \code{pec} does not support
\code{eha::aftreg} as a method for calculation of IPCWs. The function
is loosely based on \code{pec:::ipcw.cox} and only supports calculating
IPCWs at specified times (\code{what = "IPCW.times"}; see \code{what} argument
in \code{?pec::ipcw}).
}
