\name{KL}
\alias{KL}
\title{Compute symmetric version of Kullback-Leibler (KL) divergence between two categorical distributions}
\usage{
KL(x, y)
}
\arguments{
  \item{x}{The vector of probabilities in the first
  distribution}

  \item{y}{The vector of probabilities in the second
  distribution}
}
\description{
Compute symmetric version of Kullback-Leibler (KL)
divergence between two categorical distributions
}

