% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-survival-data.R
\name{label_intervals}
\alias{label_intervals}
\title{Create character vector of interval labels}
\usage{
label_intervals(tstart, tend, right_closed = TRUE)
}
\arguments{
\item{tstart}{Vector of interval start times.}

\item{tend}{Vector of interval end times.}

\item{right_closed}{Logical. If \code{TRUE} (default), intervals are assumed right_closed
closed and left open. If \code{FALSE} left closed and right_closed open.}
}
\description{
Create character vector of interval labels
}
\keyword{internal}
