% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-table.R
\name{lifetable}
\alias{lifetable}
\title{Implementation of the actuarial life table method.}
\usage{
lifetable(data, formula, cut = NULL)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a model formula}

\item{cut}{the vector of timepoints to cut at}
}
\value{
A \code{\link[tibble]{tibble}} containing riskset information as well
as hazard and survival function calculated by using the actuarial life table
method.
}
\description{
Implementation of the actuarial life table method.
}
\examples{
data("Melanoma", package="MASS")
Melanoma$status <- ifelse(Melanoma$status == 1, 1, 0)
lifetable(Melanoma, Surv(time,status)~1, cut = seq(0, 6000, 1500))
}
\seealso{
split_info
}
