% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-eha-aftreg.R
\name{predictSurvProb.aftreg}
\alias{predictSurvProb.aftreg}
\title{Survival probability for eha::aftreg objects}
\usage{
\method{predictSurvProb}{aftreg}(object, newdata, times, ...)
}
\arguments{
\item{object}{A fitted model from which to extract predicted survival
probabilities}

\item{newdata}{A data frame containing predictor variable combinations for
which to compute predicted survival probabilities.}

\item{times}{A vector of times in the range of the response variable, e.g.
times when the response is a survival object, at which to return the
survival probabilities.}

\item{...}{Additional arguments that are passed on to the current method.}
}
\description{
The function is named to be compatible with \code{pec} package. However,
\code{pec::pec} currently does not allow \code{method = 'aftreg'}. Therefore, this
function is only used internally in \code{pec_cv}.
The calculation of the Survival probability is based on the formula given
in \code{?eha::aftreg}.
}
