% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-survival-data.R
\name{split_info}
\alias{split_info}
\title{Transform standard time-to-event data to interval data format.}
\usage{
split_info(
  data,
  time_var = "time",
  status_var = "status",
  id_var = "id",
  breaks = NULL,
  right_closed = TRUE,
  max_end = FALSE
)
}
\arguments{
\item{data}{Data set from which time and status variables will be extracted.}

\item{time_var}{The variable storing event times.}

\item{status_var}{The variable storing the event indicator.}

\item{id_var}{The variable storing ID variable. Will be created if not
in data set.}

\item{breaks}{The time points of the interval borders.}

\item{right_closed}{Logical. If \code{TRUE} (default), intervals are assumed right_closed
closed and left open. If \code{FALSE} left closed and right_closed open.}

\item{max_end}{logical. Should the last interval span until the last
observed censoring or event time (if larger than the largest specified
cut point).}
}
\description{
Given a data set in standard format with one row per observation unit,
transform the data set into interval data, where intervals are specified
via breaks. Each observation unit will have as many rows as interval
start points that it survived.
}
\examples{
data(ovarian, package="survival")
split_info(ovarian, futime, fustat, breaks = c(0, 59, 500, 1500))
}
