% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-survival-data.R
\name{survived_breaks}
\alias{survived_breaks}
\title{Create index of breaks survived}
\usage{
survived_breaks(time, breaks, right_closed = TRUE)
}
\arguments{
\item{time}{The time of the event.}

\item{breaks}{The time points of the interval borders.}

\item{right_closed}{Logical. If \code{TRUE} (default), intervals are assumed right_closed
closed and left open. If \code{FALSE} left closed and right_closed open.}
}
\description{
Create index of breaks survived
}
\keyword{internal}
