% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-parser.R
\name{create_formulas}
\alias{create_formulas}
\title{Create formulas for all possible combinations of variables}
\usage{
create_formulas(covars, lhs = "Surv(time, status)~", fname = NULL)
}
\arguments{
\item{covars}{A character vector of variables for which all possible
unique model formulas will be created.}

\item{lhs}{The left-hand side of the formula to be produced.}

\item{fname}{A string of the special function in which \code{x} will be
wrapped in. Defaults to \code{"rcs"}}
}
\value{
A vector of formula strings for possible (and correct) model
specifications given the variables provided.
}
\description{
Creates formulas for all possible combinations of variables provided to
the function. All terms can appear as linear terms or non-linear terms.
}
