% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_utility.R
\name{get_coxsnell}
\alias{get_coxsnell}
\title{Extract Cox-Snell residuals from a \code{coxph} object}
\usage{
get_coxsnell(fit)
}
\arguments{
\item{fit}{
    the result of fitting a Cox regression model, using the \code{coxph} function. 
  }
}
\value{
A vector of Cox-Snell residuals.
}
\description{
Extract Cox-Snell residuals from a \code{coxph} object
}

