% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_utility.R
\name{get_coxsnell}
\alias{get_coxsnell}
\title{Extract Cox-Snell residuals from a \code{coxph} object}
\usage{
get_coxsnell(object)
}
\arguments{
\item{object}{An object of class \code{link[survival]{coxph}}}
}
\description{
Extract Cox-Snell residuals from a \code{coxph} object
}

